requirements.txt
openpyxl>=3.1.5
pillow>=10.0.0

